// src/server/api/models/position/models/Bond.ts
export type Bond = {
    /**
     * The unique ID for the record. This is the ID that is used in all services and related records when referencing the record.
     */
    id: UUID;
    /**
     * The ID of the user that created the record.
     */
    createdBy: string;
    /**
     * When the record was created.
     */
    createdAt: LocalDateTime;
    /**
     * The ID of the user that last modified the record.
     */
    updatedBy: string;
    /**
     * When the record was last modified.
     */
    updatedAt: LocalDateTime;
    /**
     * Indicates if this is a deleted record
     */
    deleted: boolean;
    /**
     * The date the instrument was issued
     */
    issueDate: LocalDate;
    /**
     * The unique security identifier for the instrument in the market where it is traded.
     */
    secId: string | null;
    /**
     * The international security identification number for the instrument.
     */
    isin?: string | null;
    /**
     * The description of the security
     */
    secDesc: string;
    /**
     * The description of the security
     */
    secNotes?: string | null;
    /**
     * The url to the image for the instrument
     */
    imageUrl?: string | null;
    /**
     * The number of days it takes to settle trades
     */
    settlementDays: number;
    /**
     * The status of the instrument
     */
    status: InstrumentStatus;
    /**
     * The type of the instrument
     */
    type: InstrumentType;
    /**
     * Specifies if you can short the instrument. The default is false
     */
    allowShorts: boolean;
    /**
     * The logo for the equity.
     */
    logo?: Blob | null;
    /**
     * The logo file name
     */
    logoFileName?: string | null;
    /**
     * The logo filename.
     */
    logoMimeType?: string | null;
    /**
     * The instrument price
     */
    price: number;
    /**
     * The instrument currency
     */
    currency: string;
    /**
     * The par value of the instrument. This is typically 1 for equity and 1000 for bonds.
     */
    parValue: number;
    /**
     *
     * The price conversion ratio applicable to the instrument. This is typically 1 for equity and 10 for bonds.
     *
     */
    priceConvRatio: number;
    marketId: UUID;
    marketCode: string;
    /**
     * The maturity date of the bond
     */
    maturityDate: LocalDate;
    /**
     * The coupon rate specified with an annualized basis.
     */
    couponRate: number;
    /**
     * The issuers full name
     */
    issuersFullName: string;
    /**
     * The minimum order value.
     */
    minOrderValue: number;
    /**
     * The maximum order value.
     */
    maxOrderValue: number;
    /**
     * The order size multiple.
     */
    orderSizeStep: number;
    /**
     * The coupon frequency of the bond
     */
    couponFrequency: BondCouponFrequency;
    /**
     * The classification of the issuer
     */
    issuerClassification: BondIssuerClassification;
    /**
     * The classification of the risk
     */
    riskClassification?: BondRiskClassification | null;
    /**
     * The coupon type of the bond
     */
    couponType: BondCouponType;
    /**
     * The coupon date generation
     */
    couponDateGeneration: BondCouponDateGeneration;
    /**
     * The coupon interest basis
     */
    couponInterestBasis: InterestBasis;
};